gl: Properly handle RGB24 updates, taking into account that the backbuffer contents are undefined after a swap. See #321 for details.
This patch has a known issue: exposing a previously offscreen window shows corrupted image for the part of the window that was previously offscreen. This is because we can't rely on the backbuffer's contents to remain there when a window is taken offscreen. Solution will require the use of framebuffer objects.